requirements.txt
numpy
torch
torchvision
opencv-python
pandas
Pillow
scikit-image
filterpy
matplotlib
